requirements.txt
fastapi
uvicorn
pydantic
openai
python-multipart
python-dotenv
sqlalchemy
alembic
